+++ backend/bracket/sql/tournaments.py
@@ -1,19 +1,7 @@
 from typing import Any
 
 from bracket.database import database
-from bracket.logic.tournaments import delete_tournament_logo
 from bracket.models.db.tournament import Tournament
-from bracket.sql.courts import sql_delete_courts_of_tournament
-from bracket.sql.players import sql_delete_players_of_tournament
-from bracket.sql.shared import (
-    sql_delete_stage_item_relations,
-)
-from bracket.sql.stage_items import sql_delete_stage_item
-from bracket.sql.stages import get_full_tournament_details, sql_delete_stage
-from bracket.sql.teams import (
-    sql_delete_teams_of_tournament,
-)
-from bracket.utils.types import assert_some
 
 
 async def sql_get_tournament(tournament_id: int) -> Tournament:
@@ -63,23 +51,3 @@ async def sql_delete_tournament(tournament_id: int) -> None:
         WHERE id = :tournament_id
         """
     await database.fetch_one(query=query, values={"tournament_id": tournament_id})
-
-
-async def sql_delete_tournament_completely(tournament_id: int) -> None:
-    stages = await get_full_tournament_details(tournament_id)
-    await delete_tournament_logo(tournament_id)
-
-    for stage in stages:
-        for stage_item in stage.stage_items:
-            await sql_delete_stage_item_relations(stage_item.id)
-
-    for stage in stages:
-        for stage_item in stage.stage_items:
-            await sql_delete_stage_item(stage_item.id)
-
-        await sql_delete_stage(tournament_id, assert_some(stage.id))
-
-    await sql_delete_players_of_tournament(tournament_id)
-    await sql_delete_courts_of_tournament(tournament_id)
-    await sql_delete_teams_of_tournament(tournament_id)
-    await sql_delete_tournament(tournament_id)
